requirements.txt
functools
joblib==1.1.0
numpy==1.23.0
openpyxl==3.0.10
pandas==1.4.3
pandas-datareader
plotly==5.9.0
pymongo==4.1.1
pymongo[srv]
PyPortfolioOpt
sklearn==0.0
statsmodels==0.13.2
streamlit==1.11.0
XlsxWriter==3.0.3
tensorflow-cpu==2.9.1